+++ tests/schema_registry/_sync/test_avro_serdes.py
@@ -38,7 +38,7 @@ from confluent_kafka.schema_registry.rules.encryption.azurekms.azure_driver impo
 from confluent_kafka.schema_registry.rules.encryption.dek_registry.dek_registry_client import \
     DekRegistryClient, DekAlgorithm
 from confluent_kafka.schema_registry.rules.encryption.encrypt_executor import \
-    FieldEncryptionExecutor, Clock
+    FieldEncryptionExecutor, Clock, EncryptionExecutor
 from confluent_kafka.schema_registry.rules.encryption.gcpkms.gcp_driver import \
     GcpKmsDriver
 from confluent_kafka.schema_registry.rules.encryption.hcvault.hcvault_driver import \
@@ -1125,7 +1125,7 @@ def test_avro_encryption():
         'bytesField': b'foobar',
     }
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client = executor.client
+    dek_client = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1133,6 +1133,130 @@ def test_avro_encryption():
     assert obj['stringField'] != 'hi'
     obj['stringField'] = 'hi'
     obj['bytesField'] = b'foobar'
+
+    deser = AvroDeserializer(client, rule_conf=rule_conf)
+    executor.executor.client = dek_client
+    obj2 = deser(obj_bytes, ser_ctx)
+    assert obj == obj2
+
+
+def test_avro_payload_encryption():
+    executor = EncryptionExecutor.register_with_clock(FakeClock())
+
+    conf = {'url': _BASE_URL}
+    client = SchemaRegistryClient.new_client(conf)
+    ser_conf = {'auto.register.schemas': False, 'use.latest.version': True}
+    rule_conf = {'secret': 'mysecret'}
+    schema = {
+        'type': 'record',
+        'name': 'test',
+        'fields': [
+            {'name': 'intField', 'type': 'int'},
+            {'name': 'doubleField', 'type': 'double'},
+            {'name': 'stringField', 'type': 'string', 'confluent:tags': ['PII']},
+            {'name': 'booleanField', 'type': 'boolean'},
+            {'name': 'bytesField', 'type': 'bytes', 'confluent:tags': ['PII']},
+        ]
+    }
+
+    rule = Rule(
+        "test-encrypt",
+        "",
+        RuleKind.TRANSFORM,
+        RuleMode.WRITEREAD,
+        "ENCRYPT_PAYLOAD",
+        None,
+        RuleParams({
+            "encrypt.kek.name": "kek1",
+            "encrypt.kms.type": "local-kms",
+            "encrypt.kms.key.id": "mykey"
+        }),
+        None,
+        None,
+        "ERROR,NONE",
+        False
+    )
+    client.register_schema(_SUBJECT, Schema(
+        json.dumps(schema),
+        "AVRO",
+        [],
+        None,
+        RuleSet(None, None, [rule])
+    ))
+
+    obj = {
+        'intField': 123,
+        'doubleField': 45.67,
+        'stringField': 'hi',
+        'booleanField': True,
+        'bytesField': b'foobar',
+    }
+    ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
+    dek_client = executor.client
+    ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
+    obj_bytes = ser(obj, ser_ctx)
+
+    deser = AvroDeserializer(client, rule_conf=rule_conf)
+    executor.client = dek_client
+    obj2 = deser(obj_bytes, ser_ctx)
+    assert obj == obj2
+
+
+def test_avro_encryption_alternate_keks():
+    executor = EncryptionExecutor.register_with_clock(FakeClock())
+
+    conf = {'url': _BASE_URL}
+    client = SchemaRegistryClient.new_client(conf)
+    ser_conf = {'auto.register.schemas': False, 'use.latest.version': True}
+    rule_conf = {'secret': 'mysecret', 'encrypt.alternate.kms.key.ids': 'mykey2,mykey3'}
+    schema = {
+        'type': 'record',
+        'name': 'test',
+        'fields': [
+            {'name': 'intField', 'type': 'int'},
+            {'name': 'doubleField', 'type': 'double'},
+            {'name': 'stringField', 'type': 'string', 'confluent:tags': ['PII']},
+            {'name': 'booleanField', 'type': 'boolean'},
+            {'name': 'bytesField', 'type': 'bytes', 'confluent:tags': ['PII']},
+        ]
+    }
+
+    rule = Rule(
+        "test-encrypt",
+        "",
+        RuleKind.TRANSFORM,
+        RuleMode.WRITEREAD,
+        "ENCRYPT_PAYLOAD",
+        None,
+        RuleParams({
+            "encrypt.kek.name": "kek1",
+            "encrypt.kms.type": "local-kms",
+            "encrypt.kms.key.id": "mykey"
+        }),
+        None,
+        None,
+        "ERROR,NONE",
+        False
+    )
+    client.register_schema(_SUBJECT, Schema(
+        json.dumps(schema),
+        "AVRO",
+        [],
+        None,
+        RuleSet(None, None, [rule])
+    ))
+
+    obj = {
+        'intField': 123,
+        'doubleField': 45.67,
+        'stringField': 'hi',
+        'booleanField': True,
+        'bytesField': b'foobar',
+    }
+    ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
+    dek_client = executor.client
+    ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
+    obj_bytes = ser(obj, ser_ctx)
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
     executor.client = dek_client
@@ -1193,7 +1317,7 @@ def test_avro_encryption_deterministic():
         'bytesField': b'foobar',
     }
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client = executor.client
+    dek_client = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1203,7 +1327,7 @@ def test_avro_encryption_deterministic():
     obj['bytesField'] = b'foobar'
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
-    executor.client = dek_client
+    executor.executor.client = dek_client
     obj2 = deser(obj_bytes, ser_ctx)
     assert obj == obj2
 
@@ -1273,7 +1397,7 @@ def test_avro_encryption_cel():
         'bytesField': b'foobar',
     }
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client = executor.client
+    dek_client = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1283,7 +1407,7 @@ def test_avro_encryption_cel():
     obj['bytesField'] = b'foobar'
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
-    executor.client = dek_client
+    executor.executor.client = dek_client
     obj2 = deser(obj_bytes, ser_ctx)
     assert obj == obj2
 
@@ -1341,7 +1465,7 @@ def test_avro_encryption_dek_rotation():
         'bytesField': b'foobar',
     }
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client: DekRegistryClient = executor.client
+    dek_client: DekRegistryClient = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1350,17 +1474,17 @@ def test_avro_encryption_dek_rotation():
     obj['stringField'] = 'hi'
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
-    executor.client = dek_client
+    executor.executor.client = dek_client
     obj2 = deser(obj_bytes, ser_ctx)
     assert obj == obj2
 
-    dek_client = executor.client
+    dek_client = executor.executor.client
     dek = dek_client.get_dek("kek1-rot", _SUBJECT, version=-1)
     assert dek.version == 1
 
     # advance 2 days
     now = datetime.now() + timedelta(days=2)
-    executor.clock.fixed_now = int(round(now.timestamp() * 1000))
+    executor.executor.clock.fixed_now = int(round(now.timestamp() * 1000))
 
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1376,7 +1500,7 @@ def test_avro_encryption_dek_rotation():
 
     # advance 2 days
     now = datetime.now() + timedelta(days=2)
-    executor.clock.fixed_now = int(round(now.timestamp() * 1000))
+    executor.executor.clock.fixed_now = int(round(now.timestamp() * 1000))
 
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1437,7 +1561,7 @@ def test_avro_encryption_f1_preserialized():
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     deser = AvroDeserializer(client, rule_conf=rule_conf)
 
-    dek_client: DekRegistryClient = executor.client
+    dek_client: DekRegistryClient = executor.executor.client
     dek_client.register_kek("kek1-f1", "local-kms", "mykey")
 
     encrypted_dek = "07V2ndh02DA73p+dTybwZFm7DKQSZN1tEwQh+FoX1DZLk4Yj2LLu4omYjp/84tAg3BYlkfGSz+zZacJHIE4="
@@ -1500,7 +1624,7 @@ def test_avro_encryption_deterministic_f1_preserialized():
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     deser = AvroDeserializer(client, rule_conf=rule_conf)
 
-    dek_client: DekRegistryClient = executor.client
+    dek_client: DekRegistryClient = executor.executor.client
     dek_client.register_kek("kek1-det-f1", "local-kms", "mykey")
 
     encrypted_dek = ("YSx3DTlAHrmpoDChquJMifmPntBzxgRVdMzgYL82rgWBKn7aUSnG+WIu9oz"
@@ -1562,7 +1686,7 @@ def test_avro_encryption_dek_rotation_f1_preserialized():
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     deser = AvroDeserializer(client, rule_conf=rule_conf)
 
-    dek_client: DekRegistryClient = executor.client
+    dek_client: DekRegistryClient = executor.executor.client
     dek_client.register_kek("kek1-rot-f1", "local-kms", "mykey")
 
     encrypted_dek = "W/v6hOQYq1idVAcs1pPWz9UUONMVZW4IrglTnG88TsWjeCjxmtRQ4VaNe/I5dCfm2zyY9Cu0nqdvqImtUk4="
@@ -1642,7 +1766,7 @@ def test_avro_encryption_references():
     ))
 
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client = executor.client
+    dek_client = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1652,7 +1776,7 @@ def test_avro_encryption_references():
     obj['refField']['bytesField'] = b'foobar'
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
-    executor.client = dek_client
+    executor.executor.client = dek_client
     obj2 = deser(obj_bytes, ser_ctx)
     assert obj == obj2
 
@@ -1709,7 +1833,7 @@ def test_avro_encryption_with_union():
         'bytesField': b'foobar',
     }
     ser = AvroSerializer(client, schema_str=None, conf=ser_conf, rule_conf=rule_conf)
-    dek_client = executor.client
+    dek_client = executor.executor.client
     ser_ctx = SerializationContext(_TOPIC, MessageField.VALUE)
     obj_bytes = ser(obj, ser_ctx)
 
@@ -1719,7 +1843,7 @@ def test_avro_encryption_with_union():
     obj['bytesField'] = b'foobar'
 
     deser = AvroDeserializer(client, rule_conf=rule_conf)
-    executor.client = dek_client
+    executor.executor.client = dek_client
     obj2 = deser(obj_bytes, ser_ctx)
     assert obj == obj2